backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.51.0
httpx==0.27.0
python-multipart==0.0.9

# GitHub Integration
PyGithub==2.1.1
gql[requests]==3.5.0
GitPython==3.1.40

# Test Optimization
numpy==1.26.0

# Test Coverage Analysis
coverage==7.4.0







